Layout structure for lexer and buffer classes.
This is not tested yet. This is a design layout for a Buffer class
which turns an iterable into something that looks like a continuous
buffer, and a lecer that uses that buffer to read a sequence of
markup, all while minimising the amount of memory allocation.  It does
this by mostly passing long text sequences as sections with is_initial
and is_final attributes, by yielding singleton objects (classes) for
the initial tests, and by creating special iterators that follow the
generator protocol, but can be reset following a StopIteration to start
iterating again.