Add InterfaceConformanceError for claimed but not provided interfaces.
Code raises TypeError for conversion of an object to an interface, where
the object does not claim to provide the interface.

Code raises InterfaceConformanceError for conversion of an object to an
interface, where the object claims to provide the interface, but does not
appear to actually implement it.

It seems useful to highlight the distinction between non-claiming and
claiming-but-not-implementing to the developer user.  But for reasons
explained in the added code comments, TypeError, AttributeError,
and NotImplementedError are not quite suitable.